Created auth context and updated navigation
