+++ package.json
@@ -14,14 +14,17 @@
     "@next/font": "13.1.6",
     "@stripe/stripe-js": "^1.46.0",
     "clsx": "^1.2.1",
+    "core-js": "^3.27.2",
+    "core-js-pure": "^3.27.2",
     "daisyui": "^2.49.0",
     "firebase": "^9.16.0",
     "next": "13.1.6",
+    "prop-types": "^15.8.1",
     "react": "18.2.0",
+    "react-avatar": "^5.0.3",
     "react-dom": "18.2.0",
     "react-query": "^3.39.3",
-    "react-spinners": "^0.13.8",
-    "react-user-avatar": "^1.10.0"
+    "react-spinners": "^0.13.8"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.13",
